Feature: Show value percentages of parent nodes or total of inputs.
Added the ability to calculate and display the percentage of
each node's value, either as the percentage of all input flows
of the node, or of all inputs of the entire diagram.

You can clearly see the difference when calculating percentages
based on parent nodes vs. total inputs when toggling the setting
on the Job Search sample diagram.